Added test option to etl script
Prepended queries that load a subset of data with 'test_' and
created similar COPY queries that load all of the data under
the original names.

Now etl.py can be run with the option 'y' to run without prompting,
thus loading all of the data, or with the option 'test' to run on a
subset of data. If no optional arguments are present, the user is
prompted to make a choice.